refactor(habitat): classify verify model by kind
