+++ tools/habitat-harness/src/service/model/verify/dto/verify.schema.ts
