+++ tools/habitat-harness/src/service/model/verify/index.ts
@@ -1,0 +1,24 @@
+/** Public verify module barrel for command orchestration and receipt contracts. */
+export type { VerifyBaseResolution, VerifyReceipt } from "./dto/verify.schema.js";
+export {
+  isVerifyReceipt,
+  stringifyVerifyReceipt,
+  VerifyBaseResolutionSchema,
+  VerifyBaseSchema,
+  VerifyCommandRecordSchema,
+  VerifyHabitatCheckSummarySchema,
+  VerifyNxAffectedSchema,
+  VerifyNxCacheTaskSchema,
+  VerifyPostStateSchema,
+  VerifyReceiptSchema,
+  validateVerifyReceipt,
+} from "./dto/verify.schema.js";
+export { resolveVerifyBaseEffect } from "./policy/base-resolution.policy.js";
+export { runAffectedVerificationEffect } from "./policy/nx-affected.policy.js";
+export { observeGitStatusEffect } from "./policy/post-state.policy.js";
+export type { VerifyReceiptInput } from "./policy/receipt.policy.js";
+export {
+  createVerifyReceipt,
+  readVerifyTargetPlan,
+  verifyAffectedTargets,
+} from "./policy/receipt.policy.js";
