--- tools/habitat-harness/src/service/model/verify/proof/index.ts
@@ -1,20 +1,0 @@
-/** Public verify module barrel for command orchestration and receipt contracts. */
-export { resolveVerifyBaseEffect } from "./base.js";
-export { runAffectedVerificationEffect } from "./nx-affected.js";
-export { observeGitStatusEffect } from "./post-state.js";
-export type { VerifyReceiptInput } from "./receipt.js";
-export { createVerifyReceipt, readVerifyTargetPlan, verifyAffectedTargets } from "./receipt.js";
-export type { VerifyBaseResolution, VerifyReceipt } from "./schema.js";
-export {
-  isVerifyReceipt,
-  stringifyVerifyReceipt,
-  VerifyBaseResolutionSchema,
-  VerifyBaseSchema,
-  VerifyCommandRecordSchema,
-  VerifyHabitatCheckSummarySchema,
-  VerifyNxAffectedSchema,
-  VerifyNxCacheTaskSchema,
-  VerifyPostStateSchema,
-  VerifyReceiptSchema,
-  validateVerifyReceipt,
-} from "./schema.js";
